DDD Part 5: ConcertEntity
Closely allign Concert Draft Creation with accepting, so that they can only be done together

+++ api/Concertable.Infrastructure/Migrations/20260415141212_InitialCreate.cs
@@ -671,7 +671,7 @@ namespace Concertable.Infrastructure.Migrations
                     Id = table.Column<int>(type: "int", nullable: false)
                         .Annotation("SqlServer:Identity", "1, 1"),
                     TicketId = table.Column<Guid>(type: "uniqueidentifier", nullable: false),
-                    Stars = table.Column<int>(type: "int", nullable: false),
+                    Stars = table.Column<byte>(type: "tinyint", nullable: false),
                     Details = table.Column<string>(type: "nvarchar(max)", nullable: true)
                 },
                 constraints: table =>
